Refactoring
- Moved allowed value class
